Fix SQL close all connections
