+++ src/fitm/model/Student.java
@@ -30,6 +30,12 @@ public class Student extends User {
             }
         } catch (SQLException ex) {
             Logger.getLogger(Course.class.getName()).log(Level.SEVERE, null, ex);
+        } finally {
+            try {
+                helper.closeResultSet(rs);
+            } catch (SQLException ex) {
+                Logger.getLogger(User.class.getName()).log(Level.SEVERE, null, ex);
+            }
         }
         return studentArrayList;
     }
