+++ src/fitm/model/User.java
@@ -52,8 +52,41 @@ public class User {
                 }
             } catch (SQLException ex) {
                 Logger.getLogger(User.class.getName()).log(Level.SEVERE, null, ex);
+            } finally {
+                try {
+                    helper.closeResultSet(rs);
+                } catch (SQLException ex) {
+                    Logger.getLogger(User.class.getName()).log(Level.SEVERE, null, ex);
+                }
             }
         }
         return flag;
+    }
+
+    public static User getUserById(String userid) throws ServletException {
+        SQLHelper helper = SQLHelper.getInstance();
+        String selection = SQLHelper.Columns.USER_ID + "=? ";
+        String[] selectionArgs = {userid};
+        User user = null;
+
+        ResultSet rs = helper.query(SQLHelper.TABLE_USER_WEB, null, selection, selectionArgs, null);
+        if (rs != null) {
+            try {
+                while (rs.next()) {
+                    user = new User(userid,
+                            rs.getString(SQLHelper.Columns.REALNAME),
+                            rs.getInt(SQLHelper.Columns.USERTYPE));
+                }
+            } catch(SQLException ex) {
+                Logger.getLogger(User.class.getName()).log(Level.SEVERE, null, ex);
+            } finally {
+                try {
+                    helper.closeResultSet(rs);
+                } catch (SQLException ex) {
+                    Logger.getLogger(User.class.getName()).log(Level.SEVERE, null, ex);
+                }
+            }
+        }
+        return user;
     }
 }
